✨ feat(board-list.tsx): add BoardList component to display a list of boards based on orgId and query parameters
The BoardList component is responsible for rendering a list of boards based on the provided orgId and query parameters. It receives the orgId and query as props.

If there is no data available and a search query is provided, it displays a message suggesting to try searching for something else.

If there is no data available and the favorites query parameter is set to true, it displays a message indicating that there are no favorites.

If there is no data available, it displays a message indicating that there are no boards.

The component returns a div element containing the JSON representation of the query.